Diagnostics
Tracking diagnostics changes for toggle function
