+++ lua/mapping/init.lua
@@ -7,14 +7,14 @@ require("mapping.basis")
 -- GIT MANAGEMENT
 require("mapping.git")
 
--- LSP
-require("mapping.lsp")
-
 -- CUSTOM
 require("mapping.custom")
 
 -- NAVIGATION THROUGH VIM
 require("mapping.navigation")
+
+-- DIAGNOSTICS
+require("mapping.diagnostics")
 
 -- DEPENDENCIES MANAGEMENT
 require("mapping.packageManager")
